@@ -1,73 +1,75 @@
 {
-  "name": "infi",
+  "name": "infi-pwa",
   "version": "0.0.0",
-  "license": "MIT",
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
     "build": "ng build",
     "test": "ng test",
     "lint": "ng lint",
-    "e2e": "ng e2e"
+    "e2e": "ng e2e",
+    "build-prod-ngsw": "ng build -prod && node_modules/.bin/ngsw-config dist ./src/ngsw-config.json && cp node_modules/@angular/service-worker/ngsw-worker.js ./dist/ngsw-worker.js"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.7",
-    "@angular/cdk": "^5.2.3",
-    "@angular/common": "^5.0.0",
-    "@angular/compiler": "^5.2.7",
-    "@angular/core": "^5.2.7",
-    "@angular/forms": "^5.2.7",
-    "@angular/http": "^5.2.7",
-    "@angular/material": "^5.2.3",
-    "@angular/platform-browser": "^5.2.3",
-    "@angular/platform-browser-dynamic": "^5.2.7",
-    "@angular/platform-server": "^5.2.10",
-    "@angular/router": "^5.2.7",
-    "@angular/service-worker": "^5.2.10",
-    "@ng-bootstrap/ng-bootstrap": "^1.0.0-beta.5",
-    "@swimlane/ngx-charts": "^7.1.1",
-    "angular-calendar": "0.22.3",
-    "angular-draggable-droppable": "^1.1.1",
+    "@angular/animations": "^6.1.0",
+    "@angular/cdk": "^6.1.0",
+    "@angular/common": "^6.1.0",
+    "@angular/compiler": "^6.1.0",
+    "@angular/core": "^6.1.0",
+    "@angular/fire": "^5.0.2",
+    "@angular/forms": "^6.1.0",
+    "@angular/http": "^6.1.0",
+    "@angular/platform-browser": "^6.1.0",
+    "@angular/platform-browser-dynamic": "^6.1.0",
+    "@angular/platform-server": "^6.1.0",
+    "@angular/pwa": "^0.8.4",
+    "@angular/router": "^6.1.0",
+    "@angular/service-worker": "^6.1.9",
+    "@ng-bootstrap/ng-bootstrap": "^3.2.2",
+    "@swimlane/ngx-charts": "^9.0.0",
+    "angular-calendar": "^0.26.1",
+    "angular-draggable-droppable": "^4.0.0",
     "angular-md5": "^0.1.10",
-    "angular-resizable-element": "^1.2.5",
-    "angular2-carousel": "^1.0.12",
+    "angular-resizable-element": "^3.2.0",
     "core-js": "^2.5.3",
     "d3": "^4.13.0",
+    "date-fns": "^1.29.0",
+    "firebase": "^5.5.3",
     "hammerjs": "^2.0.8",
     "jquery": "^3.3.1",
     "jwt-decode": "^2.2.0",
     "moment": "^2.21.0",
-    "ng-pick-datetime": "^5.1.7",
+    "ng-pick-datetime": "^6.0.0",
     "ng-pick-datetime-moment": "^1.0.5",
     "ng-pwa-tools": "0.0.15",
     "ng-sidebar": "^7.0.1",
     "ngx-toastr": "^8.5.0",
     "npm-install-peers": "^1.2.1",
-    "rxjs": "^5.5.6",
+    "rxjs": "^6.3.3",
+    "rxjs-compat": "^6.3.3",
     "ts-md5": "^1.2.4",
     "zone.js": "^0.8.20"
   },
   "devDependencies": {
-    "@angular/cli": "^1.7.4",
-    "@angular/compiler-cli": "^5.2.10",
-    "@angular/language-service": "^5.2.10",
-    "@types/jasmine": "~2.8.6",
+    "@angular-devkit/build-angular": "~0.8.0",
+    "@angular/cli": "~6.2.3",
+    "@angular/compiler-cli": "^6.1.0",
+    "@angular/language-service": "^6.1.0",
+    "@types/jasmine": "~2.8.8",
     "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~9.4.6",
-    "codelyzer": "~4.2.1",
-    "http-server": "^0.11.1",
-    "jasmine-core": "~3.1.0",
+    "@types/node": "~8.9.4",
+    "codelyzer": "~4.3.0",
+    "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~2.0.0",
+    "karma": "~3.0.0",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.4.1",
-    "karma-jasmine": "~1.1.1",
+    "karma-coverage-istanbul-reporter": "~2.0.1",
+    "karma-jasmine": "~1.1.2",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "^5.4.1",
-    "ts-node": "~5.0.1",
-    "tslint": "~5.9.1",
-    "typescript": "^2.6.1"
+    "protractor": "~5.4.0",
+    "ts-node": "~7.0.0",
+    "tslint": "~5.11.0",
+    "typescript": "~2.9.2"
   }
 }
