@@ -7,23 +7,23 @@ import {
   HttpInterceptor
 } from '@angular/common/http';
 import { AuthService } from './auth.service';
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 import { HttpResponse } from '@angular/common/http';
 import { HttpErrorResponse } from '@angular/common/http';
 export class JwtInterceptor implements HttpInterceptor {
-  constructor(public auth: AuthService) {}
-intercept(request: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {    
-  return next.handle(request).do((event: HttpEvent<any>) => {
-    if (event instanceof HttpResponse) {
-      // do stuff with response if you want
-    }
-  }, (err: any) => {
-    if (err instanceof HttpErrorResponse) {
-      if (err.status === 401) {
-        this.auth.collectFailedRequest(request);
+  constructor(public auth: AuthService) { }
+  intercept(request: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
+    return next.handle(request).do((event: HttpEvent<any>) => {
+      if (event instanceof HttpResponse) {
+        // do stuff with response if you want
       }
-    }
-  });
-}
+    }, (err: any) => {
+      if (err instanceof HttpErrorResponse) {
+        if (err.status === 401) {
+          this.auth.collectFailedRequest(request);
+        }
+      }
+    });
+  }
 
 }
