@@ -3,13 +3,13 @@ import { TestBed, async, inject } from '@angular/core/testing';
 import { AuthguardGuard } from './authguard.guard';
 
 describe('AuthguardGuard', () => {
-beforeEach(() => {
-TestBed.configureTestingModule({
-providers: [AuthguardGuard]
-});
-});
+    beforeEach(() => {
+        TestBed.configureTestingModule({
+            providers: [AuthguardGuard]
+        });
+    });
 
-it('should ...', inject([AuthguardGuard], (guard: AuthguardGuard) => {
-expect(guard).toBeTruthy();
-}));
+    it('should ...', inject([AuthguardGuard], (guard: AuthguardGuard) => {
+        expect(guard).toBeTruthy();
+    }));
 });
