@@ -1,15 +1,12 @@
-import { TestBed, inject } from '@angular/core/testing';
+import { TestBed } from '@angular/core/testing';
 
 import { UserService } from './user.service';
 
 describe('UserService', () => {
-beforeEach(() => {
-TestBed.configureTestingModule({
-providers: [UserService]
-});
-});
+  beforeEach(() => TestBed.configureTestingModule({}));
 
-it('should be created', inject([UserService], (service: UserService) => {
-expect(service).toBeTruthy();
-}));
+  it('should be created', () => {
+    const service: UserService = TestBed.get(UserService);
+    expect(service).toBeTruthy();
+  });
 });
