@@ -3,16 +3,16 @@ import { Injectable } from '@angular/core';
 @Injectable()
 export class UserService {
 
-private isUserLoggedIn = false;
-private username;
+  private isUserLoggedIn = false;
+  private username;
 
-constructor() {
-this.isUserLoggedIn=false;
-}
-setUserLoggedIn(){
-this.isUserLoggedIn=true;
-}
-getUserLoggedIn(){
-return this.isUserLoggedIn;
-}
+  constructor() {
+    this.isUserLoggedIn = false;
+  }
+  setUserLoggedIn() {
+    this.isUserLoggedIn = true;
+  }
+  getUserLoggedIn() {
+    return this.isUserLoggedIn;
+  }
 }
