@@ -1,3 +1,12 @@
 export const environment = {
-production: true
+  production: true,
+  firebaseConfig: {
+    apiKey: "AIzaSyDyMRDnOIt58FcHwUbTdK_RBCoTJEF2qRs",
+    authDomain: "infipwa.firebaseapp.com",
+    databaseURL: "https://infipwa.firebaseio.com",
+    projectId: "infipwa",
+    storageBucket: "infipwa.appspot.com",
+    messagingSenderId: "765257743324"
+
+  }
 };
