@@ -1,22 +1,22 @@
 /**
-* This file includes polyfills needed by Angular and is loaded before the app.
-* You can add your own extra polyfills to this file.
-*
-* This file is divided into 2 sections:
-*   1. Browser polyfills. These are applied before loading ZoneJS and are sorted by browsers.
-*   2. Application imports. Files imported after ZoneJS that should be loaded before your main
-*      file.
-*
-* The current setup is for so-called "evergreen" browsers; the last versions of browsers that
-* automatically update themselves. This includes Safari >= 10, Chrome >= 55 (including Opera),
-* Edge >= 13 on the desktop, and iOS 10 and Chrome on mobile.
-*
-* Learn more in https://angular.io/docs/ts/latest/guide/browser-support.html
-*/
+ * This file includes polyfills needed by Angular and is loaded before the app.
+ * You can add your own extra polyfills to this file.
+ *
+ * This file is divided into 2 sections:
+ *   1. Browser polyfills. These are applied before loading ZoneJS and are sorted by browsers.
+ *   2. Application imports. Files imported after ZoneJS that should be loaded before your main
+ *      file.
+ *
+ * The current setup is for so-called "evergreen" browsers; the last versions of browsers that
+ * automatically update themselves. This includes Safari >= 10, Chrome >= 55 (including Opera),
+ * Edge >= 13 on the desktop, and iOS 10 and Chrome on mobile.
+ *
+ * Learn more in https://angular.io/docs/ts/latest/guide/browser-support.html
+ */
 
 /***************************************************************************************************
-* BROWSER POLYFILLS
-*/
+ * BROWSER POLYFILLS
+ */
 
 /** IE9, IE10 and IE11 requires all of the following polyfills. **/
 // import 'core-js/es6/symbol';
@@ -37,36 +37,44 @@
 /** IE10 and IE11 requires the following for NgClass support on SVG elements */
 // import 'classlist.js';  // Run `npm install --save classlist.js`.
 
+/** IE10 and IE11 requires the following for the Reflect API. */
+// import 'core-js/es6/reflect';
+
+
 /** Evergreen browsers require these. **/
-import 'core-js/es6/reflect';
+// Used for reflect-metadata in JIT. If you use AOT (and only Angular decorators), you can remove.
 import 'core-js/es7/reflect';
 
 
 /**
-* Required to support Web Animations `@angular/platform-browser/animations`.
-* Needed for: All but Chrome, Firefox and Opera. http://caniuse.com/#feat=web-animation
-**/
+ * Web Animations `@angular/platform-browser/animations`
+ * Only required if AnimationBuilder is used within the application and using IE/Edge or Safari.
+ * Standard animation support in Angular DOES NOT require any polyfills (as of Angular 6.0).
+ **/
 // import 'web-animations-js';  // Run `npm install --save web-animations-js`.
 
+/**
+ * By default, zone.js will patch all possible macroTask and DomEvents
+ * user can disable parts of macroTask/DomEvents patch by setting following flags
+ */
 
+ // (window as any).__Zone_disable_requestAnimationFrame = true; // disable patch requestAnimationFrame
+ // (window as any).__Zone_disable_on_property = true; // disable patch onProperty such as onclick
+ // (window as any).__zone_symbol__BLACK_LISTED_EVENTS = ['scroll', 'mousemove']; // disable patch specified eventNames
+
+ /*
+ * in IE/Edge developer tools, the addEventListener will also be wrapped by zone.js
+ * with the following flag, it will bypass `zone.js` patch for IE/Edge
+ */
+// (window as any).__Zone_enable_cross_context_check = true;
 
 /***************************************************************************************************
-* Zone JS is required by Angular itself.
-*/
+ * Zone JS is required by default for Angular itself.
+ */
 import 'zone.js/dist/zone';  // Included with Angular CLI.
 
 
 
 /***************************************************************************************************
-* APPLICATION IMPORTS
-*/
-
-/**
-* Date, currency, decimal and percent pipes.
-* Needed for: All but Chrome, Firefox, Edge, IE11 and Safari 10
-*/
-// import 'intl';  // Run `npm install --save intl`.
-/**
-* Need to import at least one locale-data with intl.
-*/
-// import 'intl/locale-data/jsonp/en';
+ * APPLICATION IMPORTS
+ */
